move UInt256 to domain
