+++ src/ets/scala/io/iohk/ethereum/ets/vm/testopcodes.scala
@@ -1,7 +1,7 @@
 package io.iohk.ethereum.ets.vm
 
 import akka.util.ByteString
-import io.iohk.ethereum.domain.Address
+import io.iohk.ethereum.domain.{Address, UInt256}
 import io.iohk.ethereum.vm._
 
 case object TestCREATE extends CreateOp {
