+++ src/it/scala/io/iohk/ethereum/txExecTest/ForksTest.scala
@@ -1,12 +1,12 @@
 package io.iohk.ethereum.txExecTest
 
 import akka.util.ByteString
-import io.iohk.ethereum.domain.{BlockchainImpl, Receipt}
+import io.iohk.ethereum.domain.{BlockchainImpl, Receipt, UInt256}
 import io.iohk.ethereum.ledger.LedgerImpl
 import io.iohk.ethereum.txExecTest.util.FixtureProvider
 import io.iohk.ethereum.utils.{BlockchainConfig, MonetaryPolicyConfig}
 import io.iohk.ethereum.validators._
-import io.iohk.ethereum.vm.{UInt256, VM}
+import io.iohk.ethereum.vm.VM
 import org.scalatest.{FlatSpec, Matchers}
 
 class ForksTest extends FlatSpec with Matchers {
