+++ src/main/scala/io/iohk/ethereum/vm/EvmConfig.scala
@@ -1,6 +1,7 @@
 package io.iohk.ethereum.vm
 
 import akka.util.ByteString
+import io.iohk.ethereum.domain.UInt256
 import io.iohk.ethereum.utils.BlockchainConfig
 
 // scalastyle:off number.of.methods
