+++ src/main/scala/io/iohk/ethereum/vm/ExecEnv.scala
@@ -1,7 +1,7 @@
 package io.iohk.ethereum.vm
 
 import akka.util.ByteString
-import io.iohk.ethereum.domain.{ Address, BlockHeader }
+import io.iohk.ethereum.domain.{Address, BlockHeader, UInt256}
 
 /** Execution environment constants of an EVM program.
   *  See section 9.3 in Yellow Paper for more detail.
