+++ src/main/scala/io/iohk/ethereum/vm/Memory.scala
@@ -1,6 +1,7 @@
 package io.iohk.ethereum.vm
 
 import akka.util.ByteString
+import io.iohk.ethereum.domain.UInt256
 import org.spongycastle.util.encoders.Hex
 
 object Memory {
