+++ src/main/scala/io/iohk/ethereum/vm/PrecompiledContracts.scala
@@ -2,7 +2,7 @@ package io.iohk.ethereum.vm
 
 import akka.util.ByteString
 import io.iohk.ethereum.crypto._
-import io.iohk.ethereum.domain.Address
+import io.iohk.ethereum.domain.{Address, UInt256}
 import io.iohk.ethereum.utils.ByteUtils
 
 import scala.util.Try
