+++ src/test/scala/io/iohk/ethereum/ObjectGenerators.scala
@@ -6,9 +6,8 @@ import java.security.SecureRandom
 import akka.util.ByteString
 import io.iohk.ethereum.mpt.HexPrefix.bytesToNibbles
 import io.iohk.ethereum.network.p2p.messages.PV63._
-import io.iohk.ethereum.vm.UInt256
 import org.scalacheck.{Arbitrary, Gen}
-import io.iohk.ethereum.domain._
+import io.iohk.ethereum.domain.{UInt256, _}
 import io.iohk.ethereum.network.p2p.messages.CommonMessages.NewBlock
 import io.iohk.ethereum.network.p2p.messages.PV62.BlockBody
