+++ src/test/scala/io/iohk/ethereum/ledger/DeleteAccountsSpec.scala
@@ -3,8 +3,7 @@ package io.iohk.ethereum.ledger
 import io.iohk.ethereum.utils.{BlockchainConfig, Config}
 import io.iohk.ethereum.Mocks
 import io.iohk.ethereum.blockchain.sync.EphemBlockchainTestSetup
-import io.iohk.ethereum.domain.{Account, Address, BlockchainImpl}
-import io.iohk.ethereum.vm.UInt256
+import io.iohk.ethereum.domain.{Account, Address, BlockchainImpl, UInt256}
 import org.scalamock.scalatest.MockFactory
 import org.scalatest.{FlatSpec, Matchers}
