+++ src/test/scala/io/iohk/ethereum/vm/MockWorldState.scala
@@ -2,7 +2,7 @@ package io.iohk.ethereum.vm
 
 import akka.util.ByteString
 import io.iohk.ethereum.crypto.kec256
-import io.iohk.ethereum.domain.{Account, Address}
+import io.iohk.ethereum.domain.{Account, Address, UInt256}
 
 object MockWorldState {
   type PS = ProgramState[MockWorldState, MockStorage]
